+++ builder/build_grafy_windows.ps1
@@ -43,6 +43,7 @@ python -m PyInstaller --onedir --noconfirm --clean --strip `
     --hidden-import objects.input_parser `
     --hidden-import objects.config `
     --hidden-import objects.units `
+    --hidden-import objects.logger `
     main_grafy.py
 
 Get-ChildItem -Path "dist/statistika_grafy/_internal" -Filter "*.dist-info" -Directory -Recurse -ErrorAction SilentlyContinue | Remove-Item -Recurse -Force -ErrorAction SilentlyContinue