+++ builder/build_statistika_linux.sh
@@ -50,6 +50,7 @@ python3 -m PyInstaller --onedir --noconfirm --clean --strip \
     --hidden-import objects.input_parser \
     --hidden-import objects.config \
     --hidden-import objects.units \
+    --hidden-import objects.logger \
     main_statistika.py
 
 # Cleanup